Use the ad's ID rather than the POS to identify it in callbacks.
We can reliably get the DOM element ID by doing:

   function(event, level, message, service, slot, reference) {
       var id = slot.getSlotId().getDomId();
   }

The POS seems to only be accessible via one letter variables on the slot object. It used to be slot.d.pos[0] but now seems to be slot.i.pos[0]

These variables may be generated by a minimization script, and don't seem to be stable.